[FEAT] Add archive and single templates for 'Custom Chambres' plugin
Two new templates are added to the 'Custom Chambres' plugin: an archive template and a single post template. The templates will structure the display of posts related to 'Custom Chambres'. The archive template lists all the posts while the single post template shows the full details of a selected post.